platforms: rename platforms/pami to platforms/pf-pami
To keep module name consistency.
