+++ platforms/pf-pami/LxMotor.cpp
