+++ platforms/pf-pami/Makefile
