+++ platforms/pf-pami/Makefile.dep
@@ -3,7 +3,6 @@ USEMODULE += event
 USEMODULE += periph_gpio_irq
 USEMODULE += motor_driver
 USEMODULE += pca9685
-USEMODULE += pcf8575
 USEMODULE += printf_float
 USEMODULE += ps
 USEMODULE += scanf_float
