+++ platforms/pf-pami/Motor.cpp
