+++ platforms/pf-pami/include/actuators_motors_params.hpp
@@ -15,8 +15,8 @@ static const motor_driver_params_t actuators_motors_params =
     .pwm_mode = PWM_LEFT,
     .pwm_frequency = 20000U,
     .pwm_resolution = 100U,
-    .brake_level = MOTOR_BRAKE_LOW,
-    .enable_level = MOTOR_ENABLE_HIGH,
+    .brake_inverted = true,
+    .enable_inverted = false,
     .nb_motors = 2,
     .motors = {
         // Lift motor
@@ -24,7 +24,6 @@ static const motor_driver_params_t actuators_motors_params =
             .pwm_channel = 2,
             .gpio_enable = GPIO_PIN(PORT_A, 8),
             .gpio_dir0 = GPIO_PIN(PORT_A, 11),
-            .gpio_dir1_or_brake = GPIO_UNDEF,
             .gpio_dir_reverse = 0,
         },
         // Ball launcher (no direction, no enable, just PWM) & conveyor (no PWM, no direction, just the enable)
@@ -32,7 +31,6 @@ static const motor_driver_params_t actuators_motors_params =
             .pwm_channel = 1,
             .gpio_enable = GPIO_PIN(PORT_B, 1),
             .gpio_dir0 = GPIO_UNDEF,
-            .gpio_dir1_or_brake = GPIO_UNDEF,
             .gpio_dir_reverse = 0,
         },
     },
