+++ platforms/pf-pami/include/motion_control.hpp
@@ -66,10 +66,15 @@ constexpr double pulse_per_degree = (wheels_distance_pulse * 2 * M_PI) / 360;   
 constexpr double platform_min_speed_m_per_s = 0;  ///< Minimum speed (m/s)
 constexpr double platform_max_speed_m_per_s = 2;  ///< Maximum speed (m/s)
 constexpr double platform_max_acc_m_per_s2 = platform_max_speed_m_per_s / 2;   ///< Maximum acceleration (m/s²)
+constexpr double platform_max_dec_m_per_s2 = platform_max_acc_m_per_s2;   ///< Maximum deceleration (m/s²)
 constexpr double platform_max_acc_linear_mm_per_period2 = (
     (1000 * platform_max_acc_m_per_s2 * motion_control_thread_period_ms * motion_control_thread_period_ms) \
     / (1000 * 1000)
     );          ///< Maximum linear acceleration (mm/<motion_control_thread_period_ms>²)
+constexpr double platform_max_dec_linear_mm_per_period2 = (
+    (1000 * platform_max_dec_m_per_s2 * motion_control_thread_period_ms * motion_control_thread_period_ms) \
+    / (1000 * 1000)
+    );          ///< Maximum linear deceleration (mm/<motion_control_thread_period_ms>²)
 constexpr double platform_min_speed_linear_mm_per_period = (
     (1000 * platform_min_speed_m_per_s * motion_control_thread_period_ms) \
     / 1000);    ///< Minimum linear speed (mm/<motion_control_thread_period_ms>)
@@ -83,10 +88,15 @@ constexpr double platform_normal_speed_linear_mm_per_period = (platform_max_spee
 constexpr double platform_min_speed_deg_per_s = 0; ///< Maximum speed (deg/s)
 constexpr double platform_max_speed_deg_per_s = 720; ///< Maximum speed (deg/s)
 constexpr double platform_max_acc_deg_per_s2 = platform_max_speed_deg_per_s / 2 ;  ///< Maximum acceleration (deg/s²)
+constexpr double platform_max_dec_deg_per_s2 = platform_max_acc_deg_per_s2;  ///< Maximum deceleration (deg/s²)
 constexpr double platform_max_acc_angular_deg_per_period2 = (
     (platform_max_acc_deg_per_s2 * motion_control_thread_period_ms * motion_control_thread_period_ms) \
     / (1000 * 1000)
     );  ///< Maximum angular acceleration (deg/<motion_control_thread_period_ms>)
+constexpr double platform_max_dec_angular_deg_per_period2 = (
+    (platform_max_dec_deg_per_s2 * motion_control_thread_period_ms * motion_control_thread_period_ms) \
+    / (1000 * 1000)
+    );  ///< Maximum angular deceleration (deg/<motion_control_thread_period_ms>)
 constexpr double platform_min_speed_angular_deg_per_period = (
     (platform_min_speed_deg_per_s * motion_control_thread_period_ms) \
     / 1000);    ///< Minimum angular speed (deg/<motion_control_thread_period_ms>)
