+++ platforms/pf-pami/include/motion_motors_params.hpp
@@ -34,8 +34,8 @@ static const motor_driver_params_t motion_motors_params =
     .pwm_mode = PWM_LEFT,
     .pwm_frequency = 20000U,
     .pwm_resolution = 1000U,
-    .brake_level = MOTOR_BRAKE_HIGH,
-    .enable_level = MOTOR_ENABLE_HIGH,
+    .brake_inverted = false,
+    .enable_inverted = false,
     .nb_motors = 2,
     .motors = {
         // Left motor
@@ -43,7 +43,7 @@ static const motor_driver_params_t motion_motors_params =
             .pwm_channel = 0,
             .gpio_enable = GPIO_PIN(PORT_B, 10),
             .gpio_dir0 = GPIO_PIN(PORT_B, 2),
-            .gpio_dir1_or_brake = GPIO_PIN(PORT_B, 12),
+            .gpio_brake = GPIO_PIN(PORT_B, 12),
             .gpio_dir_reverse = 1,
         },
         // Right motor
@@ -51,7 +51,7 @@ static const motor_driver_params_t motion_motors_params =
             .pwm_channel = 1,
             .gpio_enable = GPIO_PIN(PORT_B, 10),
             .gpio_dir0 = GPIO_PIN(PORT_B, 0),
-            .gpio_dir1_or_brake = GPIO_PIN(PORT_C, 4),
+            .gpio_brake = GPIO_PIN(PORT_C, 4),
             .gpio_dir_reverse = 0,
         },
     },
