+++ platforms/pf-pami/motion_control.cpp
@@ -100,8 +100,14 @@ static void reset_speed_pids() {
 }
 
 /// PoseStraightFilter parameters.
+/// PoseStraightFilter parameters.
 static cogip::motion_control::PoseStraightFilterParameters pose_straight_filter_parameters =
-        cogip::motion_control::PoseStraightFilterParameters(angular_threshold, linear_threshold, linear_deceleration_threshold);
+        cogip::motion_control::PoseStraightFilterParameters(
+            angular_threshold,
+            linear_threshold,
+            angular_intermediate_threshold,
+            platform_max_dec_angular_deg_per_period2,
+            platform_max_dec_linear_mm_per_period2);
 /// PoseStraightFilter controller to make the robot always moves in a straight line.
 static cogip::motion_control::PoseStraightFilter pose_straight_filter =
         cogip::motion_control::PoseStraightFilter(&pose_straight_filter_parameters);
